assigment 5a done

+++ src/models/media-model.mjs
@@ -52,17 +52,39 @@ const addMedia = async (media) => {
 };
 
 
-const deleteMediaItemById = async (id) => {
+const deleteMediaItemById = async (id, user_id) => {
   try {
-    const sql = `DELETE FROM MediaItems WHERE media_id = ?; `;
-    const params = id;
+    const sql = `DELETE FROM MediaItems WHERE media_id = ? AND user_id = ?; `;
+    const params = [id, user_id];
     const [rows] = await promisePool.query(sql, params);
     console.log('rows',rows);
-    return rows.affectedRows;
+    return rows;
   } catch (e) {
     console.error('error', e.message);
     return {error: e.message};
   }
 }
 
-export { fetchAllMedia, fetchMediaById, deleteMediaItemById};
+
+const putMedia = async (media) => {
+  // oletus että media:lla on nää kaikki 
+  const {media_id, user_id, filename, size, mimetype, title, description} = media;
+  const sql = `UPDATE mediaItems 
+    SET filename = COALESCE(NULLIF(?, ''), filename), 
+    filesize = COALESCE(NULLIF(?, ''), filesize), media_type = COALESCE(NULLIF(?, ''), media_type), 
+    title= COALESCE(NULLIF(?, ''), title), description = COALESCE(?, description)
+    WHERE media_id = ? AND user_id = ?`;
+	// laittaa järjestyksessä (? ? ? ?...)
+  const params = [filename, size, mimetype, title, description, media_id, user_id];
+  try {
+    const rows = await promisePool.query(sql, params);
+    console.log('rows', rows);
+    // return {media_id: rows[0].insertId};
+    return rows;
+  } catch (e) {
+    console.error('error', e.message);
+    return {error: e.message};
+  }
+}
+
+export { fetchAllMedia, fetchMediaById, deleteMediaItemById, addMedia, putMedia};
